Create Thymio_balancing_on_a_ball.aesl
Thymio II balancing on a ball
A ball is a surface with a variable slope where a robot can navigate.  Thymio II can use the accelerometers to determine if it is horizontal on the top of the ball or if it is on a lateral slope. If it is on a slope, the accelerometers allow to know in which direction it needs to go to reach the top again. The lateral accelerometer is used to rotate, the one in the axis of the robot front direction to move forward and backwards.
http://wiki.thymio.org/en:thymioballeinverse